+++ src/lib.rs
@@ -71,9 +71,21 @@ pub enum RiscvChip {
     CH58X = 0x07,
     /// CH32V003 RISC-V2A series
     CH32V003 = 0x09,
+    // It should be CH643, guess from the WCH-LinkUtility UI
+    _Unkown0A = 0x0A,
+    // It should be CH32X035, need more verification
+    _Unkown0B = 0x0B,
 }
 
 impl RiscvChip {
+    /// Support flash protect commands, and info query commands
+    pub(crate) fn support_flash_protect(&self) -> bool {
+        // type 3, 2, 7, 0x0A, 0x0b do not support write protect
+        matches!(
+            self,
+            RiscvChip::CH32V103 | RiscvChip::CH32V003 | RiscvChip::CH32V20X | RiscvChip::CH32V30X
+        )
+    }
     fn can_disable_debug(&self) -> bool {
         matches!(self, RiscvChip::CH57X | RiscvChip::CH56X | RiscvChip::CH58X)
     }
@@ -94,6 +106,7 @@ impl RiscvChip {
             RiscvChip::CH32V20X => &flash_op::CH32V307,
             RiscvChip::CH32V30X => &flash_op::CH32V307,
             RiscvChip::CH58X => &flash_op::CH573,
+            _ => unimplemented!(),
         }
     }
     fn try_from_u8(value: u8) -> Result<Self> {