+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support WCH-LinkW, a CH32V208 flasher with wireless connection
+- Support WCH-Link firmware 2.9, some raw commands are changed
+- Support Flash protect and unprotect (#14)
 
 ### Fixed